Unsuccessful attempt to adjust background color of checkbox and button
I want the transparent background-color to be applied to the parent div only, not the checkbox or button, because the opacity of the div adds to that of the controls. In any case, I can't get the hover style to work in either one, even though the background looks right.

See https://stackoverflow.com/questions/4411306/transition-of-background-color